refactor: ScoreCalculationModal을 SharedModals로 통합
