+++ packages/shared/src/components/ModalContainer/index.tsx
@@ -1,13 +1,12 @@
 'use client';
 
-import { AdminModals, ClientModals, ScoreCalculationModal, SharedModals } from 'shared/components';
+import { AdminModals, ClientModals, SharedModals } from 'shared/components';
 
 const ModalContainer = () => {
   return (
     <>
       <AdminModals />
       <ClientModals />
-      <ScoreCalculationModal />
       <SharedModals />
     </>
   );
